src/thinking/websocket/ai-plugin/requirements.txt
torch>=2.0.0
transformers>=4.30.0
websockets>=11.0.3
Pillow>=9.5.0
numpy>=1.24.0 